Refactor OWID data collection
